requirements_clean.txt
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
lxml==5.4.0
MarkupSafe==3.0.2
openai==1.82.1
pillow==11.2.1
plyer==2.1.0
pydantic==2.11.5
pydantic_core==2.33.2
pystray==0.19.5
python-docx==1.1.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tkcalendar==1.6.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3